+++ tests/testthat/test-intervals.R
@@ -1,5 +1,25 @@
 context("Intervals")
 
+test_that("interval() returns zero-length vector", {
+  x <- interval()
+  expect_s4_class(x, "Interval")
+  expect_length(x, 0)
+  expect_equal(format(x), character())
+  expect_output(print(x), "<Interval[0]>", fixed = TRUE)
+
+  x <- interval(POSIXct(), POSIXct())
+  expect_s4_class(x, "Interval")
+  expect_length(x, 0)
+
+  x <- interval(POSIXct())
+  expect_s4_class(x, "Interval")
+  expect_length(x, 0)
+
+  x <- interval(NULL, POSIXct())
+  expect_s4_class(x, "Interval")
+  expect_length(x, 0)
+})
+
 test_that("is.interval works as expected", {
   expect_false(is.interval(234))
   expect_false(is.interval(as.POSIXct("2008-08-03 13:01:59", tz = "UTC")))
@@ -159,11 +179,6 @@ test_that("[<- can subset intervals", {
   expect_equal(ints[1, 1], int2)
   expect_equal(ints[2, 1], my_int)
 
-})
-
-test_that("format.Interval correctly displays intervals of length 0", {
-  int <- interval(ymd(18800101), ymd(18810101))
-  expect_output(print(int[FALSE]), "Interval\\(0)")
 })
 
 test_that("interval handles correctly time zones of Date objects", {